Combine tests
QueryBuilderImplTest was using GraphQLQueryBuilder anyway and it tested the logic more extensively while GraphQLQueryBuilderTest just ensured that the responses were null. With this change it makes more sense to call the old QueryBuilderImplTest the GraphQLQueryBuilderTest
